refactor: introduce rpcApp struct to replace closure captures in runRPC (T010)
- Create cmd/ai/rpc_app.go with rpcApp struct holding all closure-captured state
- rpcApp methods: initHelpers, initEventEmitter, registerHandlers, startDebugServer,
  emitStartEvent, buildSystemPrompt, createBaseContext, expandSkillCommands,
  compactBeforeRequest, updateCheckpointManager, parseJSONArgs, handleEvent,
  and all handler methods (handlePrompt, handleSteer, handleFollowUp, handleAbort,
  handleNewSession, handleResume, handleCompact, handleSet, handleModelSet, etc.)
- runRPC now constructs rpcApp, calls initHelpers/registerHandlers/initEventEmitter
- Mutable state fields protected by sync.Mutex: isStreaming, isCompacting,
  currentThinkingLevel, autoCompactionEnabled, steeringMode, followUpMode, etc.
- No behavior change: same RPC server, same handlers, same events
- post-runRPC functions (getWorkflowStatus, formatMessagesForDisplay) unchanged
