+++ cmd/ai/rpc_handlers.go
@@ -1,25 +1,20 @@
 package main
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	agentctx "github.com/tiancaiamao/ai/pkg/context"
 	"io"
-	"net/http"
 	_ "net/http/pprof"
 	"os"
 	"path/filepath"
 	"regexp"
-	"strconv"
 	"strings"
-	"sync"
 	"time"
 
 	"log/slog"
 
 	"github.com/tiancaiamao/ai/pkg/agent"
-	"github.com/tiancaiamao/ai/pkg/command"
 	"github.com/tiancaiamao/ai/pkg/compact"
 	"github.com/tiancaiamao/ai/pkg/config"
 	"github.com/tiancaiamao/ai/pkg/llm"
@@ -28,7 +23,6 @@ import (
 	"github.com/tiancaiamao/ai/pkg/session"
 	"github.com/tiancaiamao/ai/pkg/skill"
 	"github.com/tiancaiamao/ai/pkg/tools"
-	traceevent "github.com/tiancaiamao/ai/pkg/traceevent"
 )
 
 func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Writer, customSystemPrompt string, maxTurns int, timeout time.Duration) error {
@@ -243,149 +237,57 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 
 	slog.Info("Loaded  skills", "count", len(skillResult.Skills))
 	for _, s := range skillResult.Skills {
-				slog.Info("Skill", "name", s.Name, "description", s.Description)
+		slog.Info("Skill", "name", s.Name, "description", s.Description)
 	}
 
 	// Load skill usage stats and register find_skill tool
 	skillStats := skill.LoadStats(filepath.Join(agentDir, "skill-stats.json"))
 	registry.Register(tools.NewFindSkillTool(skillResult.Skills, skillStats))
 
-	// Build the full system prompt (used for agent and compactor)
-	buildSystemPrompt := func(currentSess *session.Session) string {
-		// Use custom system prompt if provided (e.g., via --system-prompt flag)
-		if customSystemPrompt != "" {
-			slog.Info("Using custom system prompt", "length", len(customSystemPrompt))
-			return customSystemPrompt
-		}
-		// Use workspace to get dynamic cwd for each prompt build
-		promptBuilder := prompt.NewBuilderWithWorkspace("", ws)
-				promptBuilder.SetTools(registry.All()).SetSkills(skillResult.Skills).SetSkillStats(skillStats)
-
-		// Set llm context for system prompt explanation (tells LLM about the mechanism)
-		// The actual content is injected dynamically in the agent loop.
-		if currentSess != nil {
-			sessionDir := currentSess.GetDir()
-			if sessionDir != "" {
-				wm := agentctx.NewLLMContext(sessionDir)
-				_ = wm // LLM context loaded by compact system
-			}
-		}
-
-		return promptBuilder.Build()
-	}
-	systemPrompt := buildSystemPrompt(sess)
-
-	// Helper function to create a new agent context
-	// restoreLLMContextFromCompaction restores the llm context overview.md
-	// from the latest compaction summary on the current session branch.
-	restoreLLMContextFromCompaction := func(sess *session.Session) {
-		// Get the latest compaction summary
-		summary := sess.GetLastCompactionSummary()
-		if summary == "" {
-			// No compaction summary found, nothing to restore
-			slog.Info("[resume-on-branch] No compaction summary found, skipping llm context restore")
-			return
-		}
-
-		// Get llm context and write the summary
-		sessionDir := sess.GetDir()
-		if sessionDir == "" {
-			slog.Warn("[resume-on-branch] No session directory, cannot restore llm context")
-			return
-		}
-
-		wm := agentctx.NewLLMContext(sessionDir)
-		if err := wm.WriteContent(summary); err != nil {
-			slog.Warn("[resume-on-branch] Failed to restore llm context", "error", err)
-		} else {
-			slog.Info("[resume-on-branch] Restored llm context from compaction summary", "summary_len", len(summary))
-		}
+	// --- Construct rpcApp ---
+	app := &rpcApp{
+		customSystemPrompt:    customSystemPrompt,
+		maxTurns:              maxTurns,
+		output:                output,
+		debugAddr:             debugAddr,
+		cfg:                   cfg,
+		configPath:            configPath,
+		model:                 model,
+		apiKey:                apiKey,
+		activeSpec:            activeSpec,
+		currentModelInfo:      currentModelInfo,
+		currentContextWindow:  currentContextWindow,
+		cwd:                   cwd,
+		agentDir:              agentDir,
+		sessionPath:           sessionPath,
+		sessionMgr:            sessionMgr,
+		sess:                  sess,
+		sessionID:             sessionID,
+		sessionName:           sessionName,
+		ws:                    ws,
+		registry:              registry,
+		compactor:             compactor,
+		ctxManager:            ctxManager,
+		compactorConfig:       compactorConfig,
+		traceOutputPath:       traceOutputPath,
+		skillResult:           skillResult,
+		skillStats:            skillStats,
+		autoCompactionEnabled: compactorConfig.AutoCompact,
+		steeringMode:          "all",
+		followUpMode:          "one-at-a-time",
+		currentThinkingLevel:  "high",
+		showThinking:          true,
+		showTools:             true,
+		showPrefix:            true,
+		busyMode:              "steer",
 	}
 
-	createBaseContext := func() *agentctx.AgentContext {
-		// Rebuild system prompt from the current session so llm-context paths
-		// stay in sync after /resume, /new, /fork, and branch resume operations.
-		systemPrompt = buildSystemPrompt(sess)
-		ctx := agentctx.NewAgentContext(systemPrompt)
-		for _, tool := range registry.All() {
-			ctx.AddTool(tool)
-		}
-		// Initialize llm context and set it on session for compaction summaries
-		if sess != nil {
-			sessionDir := sess.GetDir()
-			if sessionDir != "" {
-				// LLMContext is now a string, not a file manager
-				// The file management is handled internally by the session
-				ctx.LLMContext = ""
-			}
-			// Restore conversation history from session
-			ctx.RecentMessages = sess.GetMessages()
-			// Restore agent state and messages from checkpoint (preserves trigger counters, CWD, tokens, etc.)
-			// Checkpoint messages are compacted, so prefer them over full session history to reduce token pressure.
-			if sessionDir != "" {
-				if cpInfo, err := agentctx.LoadLatestCheckpoint(sessionDir); err == nil && cpInfo != nil {
-					cpPath := filepath.Join(sessionDir, cpInfo.Path)
-					// Try loading full snapshot (messages + state) from checkpoint
-					if snapshot, err := agentctx.LoadCheckpoint(sessionDir, cpInfo); err == nil && snapshot != nil {
-						if len(snapshot.RecentMessages) > 0 {
-							sessionMsgCount := len(ctx.RecentMessages)
-							ctx.RecentMessages = snapshot.RecentMessages
-							slog.Info("Restored messages from checkpoint",
-								"checkpoint_messages", len(snapshot.RecentMessages),
-								"session_messages", sessionMsgCount,
-								"saved", sessionMsgCount-len(snapshot.RecentMessages),
-							)
-						}
-						if snapshot.AgentState != nil {
-							ctx.AgentState = snapshot.AgentState
-							if snapshot.AgentState.CurrentWorkingDir != "" {
-								if err := ws.SetCWD(snapshot.AgentState.CurrentWorkingDir); err != nil {
-									slog.Warn("Failed to restore CWD from checkpoint", "cwd", snapshot.AgentState.CurrentWorkingDir, "error", err)
-								}
-							}
-							slog.Info("Restored agent state from checkpoint",
-								"turns", snapshot.AgentState.TotalTurns,
-								"tokens", snapshot.AgentState.TokensUsed,
-								"toolCallsSince", snapshot.AgentState.ToolCallsSinceLastTrigger,
-								"cwd", snapshot.AgentState.CurrentWorkingDir,
-							)
-						}
-						// Restore LLM context from checkpoint if available
-						if snapshot.LLMContext != "" {
-							ctx.LLMContext = snapshot.LLMContext
-						}
-					} else {
-						// Fallback: load only agent state if full snapshot fails
-						if savedState, err := agentctx.LoadCheckpointAgentState(cpPath); err == nil {
-							ctx.AgentState = savedState
-							if savedState.CurrentWorkingDir != "" {
-								if err := ws.SetCWD(savedState.CurrentWorkingDir); err != nil {
-									slog.Warn("Failed to restore CWD from checkpoint", "cwd", savedState.CurrentWorkingDir, "error", err)
-								}
-							}
-							slog.Info("Restored agent state from checkpoint (no messages)",
-								"turns", savedState.TotalTurns,
-								"tokens", savedState.TokensUsed,
-								"toolCallsSince", savedState.ToolCallsSinceLastTrigger,
-								"cwd", savedState.CurrentWorkingDir,
-							)
-						}
-					}
-				}
-			}
-			// Set up persistence callback for compact operations.
-			// Compact events are appended to messages.jsonl (immutable log).
-			// OnMessagesChanged (SaveMessages full rewrite) is removed.
-			ctx.OnCompactEvent = func(detail *agentctx.CompactEventDetail) error {
-				return sess.AppendCompactEvent(detail)
-			}
-		}
-		return ctx
-	}
+	app.initHelpers()
 
-	agentCtx := createBaseContext()
+	// --- Create agent context ---
+	agentCtx := app.createBaseContext()
 
-	// Pre-config: sessionWriter, sessionComp, executor, toolOutputConfig
+	// --- Pre-config: sessionWriter, sessionComp, executor, toolOutputConfig ---
 	sessionWriter := newSessionWriter(256)
 	defer sessionWriter.Close()
 	sessionComp := &sessionCompactor{
@@ -393,6 +295,8 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 		compactor: compactor,
 		writer:    sessionWriter,
 	}
+	app.sessionWriter = sessionWriter
+	app.sessionComp = sessionComp
 
 	concurrencyConfig := cfg.Concurrency
 	if concurrencyConfig == nil {
@@ -407,6 +311,7 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 	if toolOutputConfig == nil {
 		toolOutputConfig = config.DefaultToolOutputConfig()
 	}
+	app.toolOutputConfig = toolOutputConfig
 
 	// Build LoopConfig with all settings
 	loopCfg := cfg.ToLoopConfig(
@@ -436,10 +341,13 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 		loopCfg.MaxTurns = maxTurns
 		slog.Info("Max turns limit set", "max_turns", maxTurns)
 	}
+	app.loopCfg = loopCfg
 
 	// Create agent with LoopConfig
 	ag := agent.NewAgentFromConfigWithContext(model, apiKey, agentCtx, loopCfg)
 	defer ag.Shutdown()
+	ag.SetThinkingLevel("high")
+	app.ag = ag
 
 	// Start timeout watchdog if timeout is set
 	if timeout > 0 {
@@ -451,17 +359,16 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 	}
 
 	// Initialize checkpoint manager for persistent state
-	var checkpointMgr *agent.AgentContextCheckpointManager
 	if sess != nil {
 		sessionDir := sess.GetDir()
 		if mgr, err := agent.NewAgentContextCheckpointManager(sessionDir); err != nil {
 			slog.Warn("Failed to create checkpoint manager", "error", err)
-			checkpointMgr = nil
+			app.checkpointMgr = nil
 		} else {
-			checkpointMgr = mgr
+			app.checkpointMgr = mgr
 			defer func() {
-				if checkpointMgr != nil {
-					checkpointMgr.Close()
+				if app.checkpointMgr != nil {
+					app.checkpointMgr.Close()
 				}
 			}()
 		}
@@ -471,1768 +378,54 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 	slog.Info("Concurrency control enabled", "maxConcurrentTools", concurrencyConfig.MaxConcurrentTools, "toolTimeout", concurrencyConfig.ToolTimeout)
 	slog.Info("Tool output truncation", "maxChars", toolOutputConfig.MaxChars)
 
-	setAgentContext := func(ctx *agentctx.AgentContext) {
-		ag.SetContext(ctx)
-	}
-
-	// Create RPC server
+	// --- Create RPC server ---
 	server := rpc.NewServer()
 	server.SetOutput(output)
-	stateMu := sync.Mutex{}
-	isStreaming := false
-	isCompacting := false
-	currentThinkingLevel := "high"
-	autoCompactionEnabled := compactorConfig.AutoCompact
-	steeringMode := "all"
-	followUpMode := "one-at-a-time"
-	pendingSteer := false
-	var followUpQueue []string
-	showThinking := true
-	showTools := true
-	showPrefix := true
-	busyMode := "steer"
-	ag.SetThinkingLevel(currentThinkingLevel)
+	app.server = server
 
-		// Helper function to expand /skill:name commands and record usage
-	expandSkillCommands := func(text string) string {
-		expanded := skill.ExpandCommand(text, skillResult.Skills)
-		if skill.IsSkillCommand(text) {
-			skillName := skill.ExtractSkillName(text)
-			skillStats.RecordUsage(skillName)
-			if err := skillStats.Save(); err != nil {
-				slog.Error("Failed to save skill stats", "skill", skillName, "error", err)
-			}
+	// --- Register all handlers ---
+	validToolSummaryStrategies := map[string]bool{"full": true, "summary": true, "full-then-summary": true}
+	validToolSummaryAutomations := map[string]bool{"auto": true, "manual": true, "off": true}
+	validSteeringModes := map[string]bool{"all": true, "off": true}
+	validFollowUpModes := map[string]bool{"one-at-a-time": true, "queue": true, "off": true}
+	validThinkingLevels := map[string]bool{"off": true, "low": true, "medium": true, "high": true}
+
+	app.registerHandlers(
+		validToolSummaryStrategies,
+		validToolSummaryAutomations,
+		validSteeringModes,
+		validFollowUpModes,
+		validThinkingLevels,
+	)
+
+	// --- Build skill commands list ---
+	app.skillCommands = make([]rpc.SlashCommand, 0)
+	for _, cmd := range server.ListSlashCommands() {
+		if cmd.Hidden {
+			continue
 		}
-		return expanded
+		app.skillCommands = append(app.skillCommands, rpc.SlashCommand{
+			Name:        cmd.Name,
+			Description: cmd.Description,
+		})
 	}
-
-	// Trigger automatic compaction right before a new request is executed
-	// (prompt/idle-steer), instead of during resume operations.
-	compactBeforeRequest := func(trigger string) {
-		if compactor == nil || sess == nil {
-			return
-		}
-
-		messages := ag.GetMessages()
-		if !compactor.ShouldCompactOld(messages) {
-			return
-		}
-		if !sess.CanCompact(compactor) {
-			slog.Info("Pre-request compaction skipped: session not compactable",
-				"trigger", trigger,
-				"messages", len(messages),
-				"estimatedTokens", compactor.EstimateContextTokensOld(messages))
-			return
-		}
-
-		beforeCount := len(messages)
-		compactionInfo := agent.CompactionInfo{
-			Auto:    true,
-			Before:  beforeCount,
-			Trigger: trigger,
-		}
-
-		stateMu.Lock()
-		isCompacting = true
-		stateMu.Unlock()
-		server.EmitEvent(agent.NewCompactionStartEvent(compactionInfo))
-
-		err := runDetachedTraceSpan(
-			"compaction",
-			traceevent.CategoryEvent,
-			[]traceevent.Field{
-				{Key: "source", Value: "pre_request"},
-				{Key: "auto", Value: true},
-				{Key: "trigger", Value: trigger},
-				{Key: "before_messages", Value: beforeCount},
-			},
-			func(_ context.Context, span *traceevent.Span) error {
-				result, err := sess.Compact(compactor)
-				if err != nil {
-					return err
-				}
-
-				ag.GetContext().RecentMessages = sess.GetMessages()
-				afterCount := len(ag.GetMessages())
-				compactionInfo.After = afterCount
-
-				span.AddField("after_messages", afterCount)
-				span.AddField("tokens_before", result.TokensBefore)
-				span.AddField("tokens_after", result.TokensAfter)
-				return nil
-			},
-		)
-
-		stateMu.Lock()
-		isCompacting = false
-		stateMu.Unlock()
-
-		if err != nil {
-			compactionInfo.Error = err.Error()
-			if session.IsNonActionableCompactionError(err) {
-				slog.Info("Pre-request compaction skipped", "trigger", trigger, "reason", err)
-			} else {
-				slog.Error("Pre-request compaction failed", "trigger", trigger, "error", err)
-			}
-		}
-		server.EmitEvent(agent.NewCompactionEndEvent(compactionInfo))
-	}
-
-	// Helper function to update checkpoint manager for new session
-	updateCheckpointManager := func() error {
-		stateMu.Lock()
-		defer stateMu.Unlock()
-
-		// Close old checkpoint manager if exists
-		if checkpointMgr != nil {
-			if err := checkpointMgr.Close(); err != nil {
-				slog.Warn("Failed to close old checkpoint manager", "error", err)
-			}
-		}
-
-		// Create new checkpoint manager for current session
-		if sess != nil {
-			sessionDir := sess.GetDir()
-			if sessionDir != "" {
-				mgr, err := agent.NewAgentContextCheckpointManager(sessionDir)
-				if err != nil {
-					slog.Warn("Failed to create checkpoint manager", "error", err)
-					checkpointMgr = nil
-				} else {
-					checkpointMgr = mgr
-					slog.Info("Updated checkpoint manager", "sessionDir", sessionDir)
-				}
-			} else {
-				slog.Warn("Session directory is empty, checkpoint manager not updated")
-				checkpointMgr = nil
-			}
-		}
-
-		return nil
-	}
-
-	// === Protocol command handlers ===
-
-	server.Register(rpc.CommandPrompt, func(cmd rpc.RPCCommand) (any, error) {
-		// Parse prompt data
-		var data struct {
-			Message           string            `json:"message"`
-			StreamingBehavior string            `json:"streamingBehavior"`
-			Images            []json.RawMessage `json:"images"`
-		}
-		if len(cmd.Data) > 0 {
-			if err := json.Unmarshal(cmd.Data, &data); err != nil {
-				return nil, fmt.Errorf("invalid data: %w", err)
-			}
-		}
-		message := cmd.Message
-		if message == "" {
-			message = data.Message
-		}
-		message = strings.TrimSpace(message)
-		if message == "" {
-			return nil, fmt.Errorf("empty prompt message")
-		}
-		if len(data.Images) > 0 {
-			return nil, fmt.Errorf("images are not supported in this RPC implementation")
-		}
-
-		// Expand /skill:name commands BEFORE generic slash dispatch.
-		// Skill commands like /skill:name are not registered as slash handlers;
-		// they are expanded into full prompts and processed normally.
-		if skill.IsSkillCommand(message) {
-			expandedMessage := expandSkillCommands(message)
-			slog.Info("Expanded skill command", "original", message, "skill", skill.ExtractSkillName(message))
-
-			stateMu.Lock()
-			streaming := isStreaming
-			stateMu.Unlock()
-
-			if streaming {
-				// During streaming, treat expanded skill prompt as a steer
-				stateMu.Lock()
-				pendingSteer = true
-				stateMu.Unlock()
-				ag.Steer(expandedMessage)
-				return nil, nil
-			}
-
-			compactBeforeRequest("pre_request_prompt")
-			return nil, ag.Prompt(expandedMessage)
-		}
-
-		// Intercept slash commands — execute synchronously without agent.
-		// Only non-skill slash commands (e.g. /get_state, /compact) reach here.
-		if message[0] == '/' {
-			cmdName, args, err := command.ParseSlashCommand(message)
-			if err != nil {
-				return nil, fmt.Errorf("invalid slash command: %w", err)
-			}
-			handler, ok := server.GetSlashHandler(cmdName)
-			if !ok {
-				return nil, fmt.Errorf("unknown command: /%s", cmdName)
-			}
-			return handler(args)
-		}
-
-		stateMu.Lock()
-		streaming := isStreaming
-		mode := steeringMode
-		followMode := followUpMode
-		pending := pendingSteer
-		stateMu.Unlock()
-
-		if streaming {
-			behavior := strings.TrimSpace(data.StreamingBehavior)
-			if behavior == "" {
-				behavior = "steer" // default: steer when streaming
-			}
-			switch behavior {
-			case "steer":
-				if mode == "one-at-a-time" && pending {
-					return nil, fmt.Errorf("steer already pending")
-				}
-				stateMu.Lock()
-				pendingSteer = true
-				stateMu.Unlock()
-				ag.Steer(message)
-				return nil, nil
-			case "followUp", "follow_up":
-				if followMode == "one-at-a-time" && ag.GetPendingFollowUps() > 0 {
-					return nil, fmt.Errorf("follow-up queue already has a pending message")
-				}
-				return nil, ag.FollowUp(message)
-			case "reject":
-				return nil, fmt.Errorf("agent is streaming; rejected by busy-mode policy")
-			default:
-				return nil, fmt.Errorf("invalid streamingBehavior: %s", behavior)
-			}
-		}
-
-		compactBeforeRequest("pre_request_prompt")
-		return nil, ag.Prompt(message)
-	})
-
-	server.Register(rpc.CommandSteer, func(cmd rpc.RPCCommand) (any, error) {
-		message := cmd.Message
-		if message == "" && len(cmd.Data) > 0 {
-			var data struct {
-				Message string `json:"message"`
-			}
-			if err := json.Unmarshal(cmd.Data, &data); err != nil {
-				return nil, fmt.Errorf("invalid data: %w", err)
-			}
-			message = data.Message
-		}
-
-		slog.Info("Received steer:", "value", message)
-
-		if strings.TrimSpace(message) == "" {
-			return nil, fmt.Errorf("empty steer message")
-		}
-
-		// Expand /skill:name commands
-		expandedMessage := expandSkillCommands(message)
-		if skill.IsSkillCommand(message) {
-			slog.Info("Expanded skill command in steer", "original", message, "skill", skill.ExtractSkillName(message))
-		}
-
-		stateMu.Lock()
-		mode := steeringMode
-		pending := pendingSteer
-		streaming := isStreaming
-		stateMu.Unlock()
-		if mode == "one-at-a-time" && pending {
-			return nil, fmt.Errorf("steer already pending")
-		}
-		if !streaming {
-			compactBeforeRequest("pre_request_steer")
-		}
-		stateMu.Lock()
-		pendingSteer = true
-		stateMu.Unlock()
-		ag.Steer(expandedMessage)
-		return nil, nil
-	})
-
-	server.Register(rpc.CommandFollowUp, func(cmd rpc.RPCCommand) (any, error) {
-		message := cmd.Message
-		if message == "" && len(cmd.Data) > 0 {
-			var data struct {
-				Message string `json:"message"`
-			}
-			if err := json.Unmarshal(cmd.Data, &data); err != nil {
-				return nil, fmt.Errorf("invalid data: %w", err)
-			}
-			message = data.Message
-		}
-
-		slog.Info("Received follow_up:", "value", message)
-		if strings.TrimSpace(message) == "" {
-			return nil, fmt.Errorf("empty follow-up message")
-		}
-
-		// Expand /skill:name commands
-		expandedMessage := expandSkillCommands(message)
-		if skill.IsSkillCommand(message) {
-			slog.Info("Expanded skill command in follow_up", "original", message, "skill", skill.ExtractSkillName(message))
-		}
-
-		stateMu.Lock()
-		mode := followUpMode
-		stateMu.Unlock()
-		if mode == "one-at-a-time" && ag.GetPendingFollowUps() > 0 {
-			return nil, fmt.Errorf("follow-up queue already has a pending message")
-		}
-		return nil, ag.FollowUp(expandedMessage)
-	})
-
-	server.Register(rpc.CommandAbort, func(cmd rpc.RPCCommand) (any, error) {
-		slog.Info("Received abort")
-		ag.Abort()
-		return nil, nil
-	})
-
-	// === Slash command handlers ===
-	// All commands below are registered as slash commands.
-	// They are invoked via the prompt channel: {"type": "prompt", "message": "/command args"}
-	// and intercepted by the prompt handler before entering the agent loop.
-	// For backward compatibility, they also handle direct JSON-RPC calls where
-	// args is raw JSON from cmd.Data (e.g. {"provider":"xxx","modelId":"yyy"}).
-
-	// parseJSONArgs attempts to unmarshal args as JSON into target.
-	// Returns true if args looks like JSON and was successfully parsed.
-	parseJSONArgs := func(args string, target any) bool {
-		args = strings.TrimSpace(args)
-		if len(args) > 0 && args[0] == '{' {
-			return json.Unmarshal([]byte(args), target) == nil
-		}
-		return false
-	}
-
-	// clear_session removed: never used per spec
-
-	server.RegisterSlash("new", "Create a new session and switch to it", func(args string) (any, error) {
-		var name, title string
-		var jsonData struct {
-			Name  string `json:"name"`
-			Title string `json:"title"`
-		}
-		if parseJSONArgs(args, &jsonData) {
-			name, title = jsonData.Name, jsonData.Title
-		} else {
-			parts := strings.SplitN(args, " ", 2)
-			name = strings.TrimSpace(parts[0])
-			if len(parts) > 1 {
-				title = strings.TrimSpace(parts[1])
-			}
-		}
-		slog.Info("Received new_session", "name", name, "title", title)
-		if strings.TrimSpace(name) == "" {
-			name = time.Now().Format("20060102-150405")
-		}
-		if strings.TrimSpace(title) == "" {
-			title = name
-		}
-		newSess, err := sessionMgr.CreateSession(name, title)
-		if err != nil {
-			return nil, err
-		}
-
-		newSessionID := newSess.GetID()
-
-		// Update session manager's current ID
-		if err := sessionMgr.SetCurrent(newSessionID); err != nil {
-			return nil, err
-		}
-
-		// Update current session metadata
-		if err := sessionMgr.SaveCurrent(); err != nil {
-			slog.Info("Failed to update session metadata:", "value", err)
-		}
-
-		sess = newSess
-		sessionComp.Update(sess, compactor)
-		setAgentContext(createBaseContext())
-
-		// Update checkpoint manager for new session
-		if err := updateCheckpointManager(); err != nil {
-			slog.Warn("Failed to update checkpoint manager for new session", "error", err)
-		}
-
-		stateMu.Lock()
-		sessionID = newSessionID
-		sessionName = name
-		stateMu.Unlock()
-
-		slog.Info("Created new session", "name", name, "id", newSessionID)
-		server.EmitEvent(map[string]any{"type": "session_switch", "session": newSessionID, "sessionName": name})
-		return map[string]any{"sessionId": newSessionID, "cancelled": false}, nil
-	})
-
-	// list_sessions removed: /resume (no args) lists sessions
-	// switch_session removed: /resume handles session switching
-	// delete_session removed: use scripts/clean-sessions.sh per spec
-	// set_session_name removed: now routed through /set session-name
-
-	server.RegisterSlash("session", "Get the current agent state (model, session, streaming status)", func(args string) (any, error) {
-		slog.Info("Received get_state")
-		compactionState := buildCompactionState(compactorConfig, compactor)
-		stateMu.Lock()
-		currentSessionID := sessionID
-		currentSessionName := sessionName
-		streaming := isStreaming
-		compacting := isCompacting
-		thinkingLevel := currentThinkingLevel
-		autoCompact := autoCompactionEnabled
-		currentSteeringMode := steeringMode
-		currentFollowUpMode := followUpMode
-		modelInfo := currentModelInfo
-		stateMu.Unlock()
-
-		// Get current trace file path from FileHandler
-		aiLogPath := traceOutputPath
-		if handler := traceevent.GetHandler(); handler != nil {
-			if fh, ok := handler.(*traceevent.FileHandler); ok {
-				aiLogPath = fh.TraceFilePath("")
-			}
-		}
-
-		return &rpc.SessionState{
-			Model:                 &modelInfo,
-			ThinkingLevel:         thinkingLevel,
-			IsStreaming:           streaming,
-			IsCompacting:          compacting,
-			SteeringMode:          currentSteeringMode,
-			FollowUpMode:          currentFollowUpMode,
-			SessionFile:           sess.GetPath(),
-			SessionID:             currentSessionID,
-			SessionName:           currentSessionName,
-			AIPid:                 os.Getpid(),
-			AILogPath:             aiLogPath,
-			AIWorkingDir:          ws.GetCWD(),
-			AIStartupPath:         ws.GetGitRoot(),
-			AutoCompactionEnabled: autoCompact,
-			MessageCount:          len(ag.GetMessages()),
-			PendingMessageCount:   ag.GetPendingFollowUps(),
-			Compaction:            compactionState,
-		}, nil
-	})
-
-	server.RegisterSlash("messages", "Get formatted message summaries for the current session", func(args string) (any, error) {
-		slog.Info("Received get_messages", "args", args)
-		const defaultCount = 20
-		const maxPreviewLen = 200
-
-		count := defaultCount
-		args = strings.TrimSpace(args)
-		if args != "" {
-			if n, err := strconv.Atoi(args); err == nil && n > 0 {
-				count = n
-			}
-		}
-
-		messages := ag.GetMessages()
-		return formatMessagesForDisplay(messages, count, maxPreviewLen), nil
-	})
-
-	server.RegisterSlash("compact", "Compact conversation history to reduce context size", func(args string) (any, error) {
-		slog.Info("Received compact")
-		beforeCount := len(ag.GetMessages())
-
-		// Estimate current context usage
-		estimatedTokens := compactor.EstimateContextTokensOld(ag.GetMessages())
-		keepTokens := compactor.KeepRecentTokens()
-
-		// Check if compaction is possible before starting
-		if !sess.CanCompact(compactor) {
-			if estimatedTokens < keepTokens {
-				return nil, fmt.Errorf("all %d messages (%d tokens) fit within keep-recent budget (%d tokens); no compaction needed",
-					beforeCount, estimatedTokens, keepTokens)
-			}
-			return nil, fmt.Errorf("no messages available for compaction (all within retention window)")
-		}
-
-		compactionInfo := agent.CompactionInfo{
-			Auto:    false,
-			Before:  beforeCount,
-			Trigger: "manual_command",
-		}
-		stateMu.Lock()
-		isCompacting = true
-		stateMu.Unlock()
-		server.EmitEvent(agent.NewCompactionStartEvent(compactionInfo))
-		defer func() {
-			stateMu.Lock()
-			isCompacting = false
-			stateMu.Unlock()
-			server.EmitEvent(agent.NewCompactionEndEvent(compactionInfo))
-		}()
-
-		var response *rpc.CompactResult
-		err := runDetachedTraceSpan(
-			"compaction",
-			traceevent.CategoryEvent,
-			[]traceevent.Field{{Key: "source", Value: "manual"}},
-			func(_ context.Context, span *traceevent.Span) error {
-				span.AddField("before_messages", beforeCount)
-
-				result, err := sess.Compact(compactor)
-				if err != nil {
-					slog.Info("Compact failed:", "value", err)
-					return err
-				}
-
-				// Replace messages with compacted version
-				ag.GetContext().RecentMessages = sess.GetMessages()
-
-				afterCount := len(ag.GetMessages())
-				span.AddField("after_messages", afterCount)
-				span.AddField("tokens_before", result.TokensBefore)
-				span.AddField("tokens_after", result.TokensAfter)
-				compactionInfo.After = afterCount
-
-				slog.Info("Compact successful", "before", beforeCount, "after", afterCount)
-				response = &rpc.CompactResult{
-					FirstKeptEntryID: result.FirstKeptEntryID,
-					TokensBefore:     result.TokensBefore,
-					TokensAfter:      result.TokensAfter,
-				}
-
-				// Create checkpoint after manual compaction to preserve AgentState for resume
-				if checkpointMgr != nil && checkpointMgr.ShouldCheckpoint() {
-					agentCtx := ag.GetContext()
-					slog.Info("[Loop] Creating checkpoint after manual compact", "trigger", "manual_command", "turn", agentCtx.AgentState.TotalTurns)
-					checkpointTurn, err := checkpointMgr.CreateSnapshot(agentCtx, agentCtx.LLMContext, agentCtx.AgentState.TotalTurns)
-					if err != nil {
-						slog.Warn("[Loop] Failed to create checkpoint after manual compact", "error", err, "turn", agentCtx.AgentState.TotalTurns)
-					} else {
-						slog.Info("[Loop] Checkpoint created after manual compact", "trigger", "manual_command", "checkpoint_turn", checkpointTurn)
-					}
-				}
-				return nil
-			},
-		)
-		if err != nil {
-			compactionInfo.Error = err.Error()
-			return nil, err
-		}
-		return response, nil
-	})
-
-	// get_available_models removed: /model already lists available models
-
-	server.RegisterHiddenSlash("set_model", "Set the active model by ID (internal, use /model instead)", func(args string) (any, error) {
-		var provider, modelID string
-		var jsonData struct {
-			Provider string `json:"provider"`
-			ModelID  string `json:"modelId"`
-		}
-		if parseJSONArgs(args, &jsonData) {
-			provider, modelID = jsonData.Provider, jsonData.ModelID
-		} else {
-			parts := strings.SplitN(args, " ", 2)
-			provider = strings.TrimSpace(parts[0])
-			if len(parts) > 1 {
-				modelID = strings.TrimSpace(parts[1])
-			}
-		}
-		slog.Info("Received set_model", "provider", provider, "modelId", modelID)
-		if strings.TrimSpace(provider) == "" || strings.TrimSpace(modelID) == "" {
-			return nil, fmt.Errorf("provider and modelId are required")
-		}
-
-		specs, modelsPath, err := loadModelSpecs(cfg)
-		if err != nil {
-			return nil, fmt.Errorf("load models from %s: %w", modelsPath, err)
-		}
-		filtered := filterModelSpecsWithKeys(specs)
-		spec, ok := findModelSpec(filtered, provider, modelID)
-		if !ok {
-			if _, exists := findModelSpec(specs, provider, modelID); exists {
-				authPath, _ := config.GetDefaultAuthPath()
-				envVar := strings.ToUpper(strings.TrimSpace(provider)) + "_API_KEY"
-				return nil, fmt.Errorf("no API key for %q (set %s or update %s)", provider, envVar, authPath)
-			}
-			return nil, fmt.Errorf("model not found: %s/%s (edit %s)", provider, modelID, modelsPath)
-		}
-		if strings.TrimSpace(spec.BaseURL) == "" || strings.TrimSpace(spec.API) == "" {
-			return nil, fmt.Errorf("model %s/%s missing baseUrl or api in %s", spec.Provider, spec.ID, modelsPath)
-		}
-
-		newAPIKey, err := config.ResolveAPIKey(spec.Provider)
-		if err != nil {
-			return nil, err
-		}
-
-		model = llm.Model{
-			ID:            spec.ID,
-			Provider:      spec.Provider,
-			BaseURL:       spec.BaseURL,
-			API:           spec.API,
-			ContextWindow: spec.ContextWindow,
-			MaxTokens:     spec.MaxTokens,
-		}
-		apiKey = newAPIKey
-
-		cfg.Model.ID = spec.ID
-		cfg.Model.Provider = spec.Provider
-		cfg.Model.BaseURL = spec.BaseURL
-		cfg.Model.API = spec.API
-		cfg.Model.MaxTokens = spec.MaxTokens
-
-		ag.SetModel(model)
-		ag.SetAPIKey(apiKey)
-
-		// Recreate compactor with new model
-		compactor = compact.NewCompactor(compactorConfig, model, apiKey, systemPrompt, spec.ContextWindow)
-		sessionComp.Update(sess, compactor)
-		ag.SetCompactor(sessionComp)
-		ag.SetContextWindow(spec.ContextWindow)
-
-		if err := config.SaveConfig(cfg, configPath); err != nil {
-			slog.Info("Failed to save config:", "value", err)
-		}
-
-		info := modelInfoFromSpec(spec)
-		stateMu.Lock()
-		currentModelInfo = info
-		currentContextWindow = spec.ContextWindow
-		stateMu.Unlock()
-		return &info, nil
-	})
-
-	// cycle_model removed: use /model to set model directly
-
-	skillCommands := buildSkillCommands(skillResult.Skills)
-
-	// get_commands removed: /skills replaces it (skillCommands kept for /skills handler)
-
-	// get_session_stats removed: /session covers it
-
-	// bash removed: agent should not run raw bash via slash command
-	// abort_bash removed: /abort is sufficient
-	// abort_retry removed: /abort is sufficient
-
-	server.RegisterSlash("export_html", "Export the current session as HTML", func(args string) (any, error) {
-		slog.Info("Received export_html", "outputPath", args)
-		return "", fmt.Errorf("export_html is not supported")
-	})
-
-	// Validation maps for /set subcommands.
-	validToolSummaryStrategies := map[string]bool{
-		"llm":       true,
-		"heuristic": true,
-		"off":       true,
-	}
-	validToolSummaryAutomations := map[string]bool{
-		"off":      true,
-		"fallback": true,
-		"always":   true,
-	}
-	validSteeringModes := map[string]bool{
-		"all":           true,
-		"immediate":     true,
-		"one-at-a-time": true,
-	}
-	validFollowUpModes := map[string]bool{
-		"all":           true,
-		"immediate":     true,
-		"one-at-a-time": true,
-	}
-	validThinkingLevels := map[string]bool{
-		"off":     true,
-		"minimal": true,
-		"low":     true,
-		"medium":  true,
-		"high":    true,
-		"xhigh":   true,
-	}
-
-	// /set — unified settings dispatcher.
-	// Routes subcommands to the appropriate handler logic.
-	// Usage: /set <key> [value]
-	// No args or "help" shows available settings.
-	server.RegisterSlash("set", "Configure agent settings", func(args string) (any, error) {
-		parts := strings.Fields(args)
-		if len(parts) == 0 || parts[0] == "help" {
-			return map[string]any{
-				"usage": "/set <key> [value]",
-				"settings": []string{
-					"auto-retry <on|off>",
-					"auto-compaction <on|off>",
-					"busy-mode <steer|follow-up|reject>",
-					"follow-up-mode <all|immediate|one-at-a-time>",
-					"prefix-display <on|off|toggle>",
-					"session-name <name>",
-					"steering-mode <all|immediate|one-at-a-time>",
-					"thinking-display <on|off|toggle>",
-					"thinking-level <off|minimal|low|medium|high|xhigh>",
-					"tool-call-cutoff <n>",
-					"tool-summary-automation <off|fallback|always>",
-					"tool-summary-strategy <llm|heuristic|off>",
-					"tools-display <on|off|toggle>",
-					"trace-events [on|off|all|enable <selectors>|disable <selectors>]",
-				},
-			}, nil
-		}
-
-		key := parts[0]
-		value := ""
-		if len(parts) > 1 {
-			value = strings.TrimSpace(strings.TrimPrefix(args, key))
-		}
-
-		switch key {
-		case "auto-retry":
-			enabled := strings.TrimSpace(strings.ToLower(value))
-			var jsonData struct {
-				Enabled bool `json:"enabled"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				ag.SetAutoRetry(jsonData.Enabled)
-				slog.Info("set auto-retry", "enabled", jsonData.Enabled)
-				return map[string]any{"setting": "auto-retry", "value": jsonData.Enabled}, nil
-			}
-			val := enabled == "true" || enabled == "1"
-			ag.SetAutoRetry(val)
-			slog.Info("set auto-retry", "enabled", val)
-			return map[string]any{"setting": "auto-retry", "value": val}, nil
-
-		case "auto-compaction":
-			enabled := strings.TrimSpace(strings.ToLower(value))
-			var jsonData struct {
-				Enabled bool `json:"enabled"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				compactorConfig.AutoCompact = jsonData.Enabled
-				stateMu.Lock()
-				autoCompactionEnabled = jsonData.Enabled
-				stateMu.Unlock()
-				slog.Info("set auto-compaction", "enabled", jsonData.Enabled)
-				return map[string]any{"setting": "auto-compaction", "value": jsonData.Enabled}, nil
-			}
-			val := enabled == "true" || enabled == "1"
-			compactorConfig.AutoCompact = val
-			stateMu.Lock()
-			autoCompactionEnabled = val
-			stateMu.Unlock()
-			slog.Info("set auto-compaction", "enabled", val)
-			return map[string]any{"setting": "auto-compaction", "value": val}, nil
-
-		case "busy-mode":
-			mode := strings.TrimSpace(value)
-			switch mode {
-			case "steer", "follow-up", "reject":
-				busyMode = mode
-				return map[string]any{"setting": "busy-mode", "value": busyMode}, nil
-			default:
-				return nil, fmt.Errorf("usage: /set busy-mode <steer|follow-up|reject>")
-			}
-
-		case "follow-up-mode":
-			mode := strings.ToLower(strings.TrimSpace(value))
-			var jsonData struct {
-				Mode string `json:"mode"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				mode = strings.ToLower(jsonData.Mode)
-			}
-			if !validFollowUpModes[mode] {
-				return nil, fmt.Errorf("invalid follow-up mode; valid: all, immediate, one-at-a-time")
-			}
-			stateMu.Lock()
-			followUpMode = mode
-			stateMu.Unlock()
-			return map[string]any{"setting": "follow-up-mode", "value": mode}, nil
-
-		case "prefix-display":
-			switch strings.TrimSpace(value) {
-			case "on":
-				showPrefix = true
-			case "off":
-				showPrefix = false
-			case "toggle", "":
-				showPrefix = !showPrefix
-			default:
-				return nil, fmt.Errorf("usage: /set prefix-display <on|off|toggle>")
-			}
-			return map[string]any{"setting": "prefix-display", "value": showPrefix}, nil
-
-		case "steering-mode":
-			mode := strings.ToLower(strings.TrimSpace(value))
-			var jsonData struct {
-				Mode string `json:"mode"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				mode = strings.ToLower(jsonData.Mode)
-			}
-			if !validSteeringModes[mode] {
-				return nil, fmt.Errorf("invalid steering mode; valid: all, immediate, one-at-a-time")
-			}
-			stateMu.Lock()
-			steeringMode = mode
-			stateMu.Unlock()
-			return map[string]any{"setting": "steering-mode", "value": mode}, nil
-
-		case "thinking-display":
-			switch strings.TrimSpace(value) {
-			case "on":
-				showThinking = true
-			case "off":
-				showThinking = false
-			case "toggle", "":
-				showThinking = !showThinking
-			default:
-				return nil, fmt.Errorf("usage: /set thinking-display <on|off|toggle>")
-			}
-			return map[string]any{"setting": "thinking-display", "value": showThinking}, nil
-
-		case "thinking-level":
-			level := strings.ToLower(strings.TrimSpace(value))
-			var jsonData struct {
-				Level string `json:"level"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				level = strings.ToLower(strings.TrimSpace(jsonData.Level))
-			}
-			if !validThinkingLevels[level] {
-				return nil, fmt.Errorf("invalid thinking level; valid: off, minimal, low, medium, high, xhigh")
-			}
-			stateMu.Lock()
-			currentThinkingLevel = level
-			stateMu.Unlock()
-			ag.SetThinkingLevel(level)
-			return map[string]any{"setting": "thinking-level", "value": level}, nil
-
-		case "tool-call-cutoff":
-			var cutoff int
-			var jsonData struct {
-				Cutoff int `json:"cutoff"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				cutoff = jsonData.Cutoff
-			} else {
-				var err error
-				cutoff, err = strconv.Atoi(strings.TrimSpace(value))
-				if err != nil {
-					return nil, fmt.Errorf("invalid cutoff value: %w", err)
-				}
-			}
-			if cutoff < 0 {
-				return nil, fmt.Errorf("cutoff must be >= 0")
-			}
-			compactorConfig.ToolCallCutoff = cutoff
-			ag.SetToolCallCutoff(cutoff)
-			if err := config.SaveConfig(cfg, configPath); err != nil {
-				slog.Info("Failed to save config:", "value", err)
-			}
-			return map[string]any{"setting": "tool-call-cutoff", "value": cutoff}, nil
-
-		case "tool-summary-automation":
-			var jsonData struct {
-				Mode string `json:"mode"`
-			}
-			mode := strings.ToLower(strings.TrimSpace(value))
-			if parseJSONArgs(value, &jsonData) {
-				mode = strings.ToLower(jsonData.Mode)
-			}
-			if !validToolSummaryAutomations[mode] {
-				return nil, fmt.Errorf("invalid tool summary automation mode; valid: off, fallback, always")
-			}
-			compactorConfig.ToolSummaryAutomation = mode
-			if err := config.SaveConfig(cfg, configPath); err != nil {
-				slog.Info("Failed to save config:", "value", err)
-			}
-			return map[string]any{"setting": "tool-summary-automation", "value": mode}, nil
-
-		case "tool-summary-strategy":
-			var jsonData struct {
-				Strategy string `json:"strategy"`
-			}
-			strategy := strings.ToLower(strings.TrimSpace(value))
-			if parseJSONArgs(value, &jsonData) {
-				strategy = strings.ToLower(jsonData.Strategy)
-			}
-			if !validToolSummaryStrategies[strategy] {
-				return nil, fmt.Errorf("invalid tool summary strategy; valid: llm, heuristic, off")
-			}
-			compactorConfig.ToolSummaryStrategy = strategy
-			if err := config.SaveConfig(cfg, configPath); err != nil {
-				slog.Info("Failed to save config:", "value", err)
-			}
-			return map[string]any{"setting": "tool-summary-strategy", "value": strategy}, nil
-
-		case "tools-display":
-			switch strings.TrimSpace(value) {
-			case "on":
-				showTools = true
-			case "off":
-				showTools = false
-			case "toggle", "":
-				showTools = !showTools
-			default:
-				return nil, fmt.Errorf("usage: /set tools-display <on|off|toggle>")
-			}
-			return map[string]any{"setting": "tools-display", "value": showTools}, nil
-
-		case "session-name":
-			name := strings.TrimSpace(value)
-			var jsonData struct {
-				Name string `json:"name"`
-			}
-			if parseJSONArgs(value, &jsonData) {
-				name = jsonData.Name
-			}
-			slog.Info("set session-name", "name", name)
-			if name == "" {
-				return nil, fmt.Errorf("session name cannot be empty")
-			}
-			if _, err := sess.AppendSessionInfo(name, ""); err != nil {
-				return nil, err
-			}
-			if err := sessionMgr.UpdateSessionName(sessionID, name, ""); err != nil {
-				slog.Info("Failed to update session metadata:", "value", err)
-			}
-			if err := sessionMgr.SaveCurrent(); err != nil {
-				slog.Info("Failed to update session metadata:", "value", err)
-			}
-			stateMu.Lock()
-			sessionName = name
-			stateMu.Unlock()
-			return map[string]any{"setting": "session-name", "value": name}, nil
-
-		case "trace-events":
-			// No args = show current config.
-			if strings.TrimSpace(value) == "" {
-				return map[string]any{"events": traceevent.GetEnabledEvents()}, nil
-			}
-
-			events := strings.Fields(value)
-			var jsonData struct {
-				Events []string `json:"events"`
-			}
-			if parseJSONArgs(value, &jsonData) && len(jsonData.Events) > 0 {
-				events = jsonData.Events
-			}
-			slog.Info("set trace-events", "events", events)
-
-			normalized := make([]string, 0, len(events))
-			for _, e := range events {
-				e = strings.ToLower(strings.TrimSpace(e))
-				if e == "" {
-					continue
-				}
-				normalized = append(normalized, e)
-			}
-			if len(normalized) == 0 {
-				return map[string]any{"events": traceevent.ResetToDefaultEvents()}, nil
-			}
-
-			applyExpanded := func(expanded []string, replace bool) []string {
-				if replace {
-					traceevent.DisableAllEvents()
-				}
-				for _, eventName := range expanded {
-					traceevent.EnableEvent(eventName)
-				}
-				return traceevent.GetEnabledEvents()
-			}
-
-			op := normalized[0]
-			switch op {
-			case "on":
-				return map[string]any{"events": traceevent.ResetToDefaultEvents()}, nil
-			case "all":
-				expanded, _ := traceevent.ExpandEventSelectors([]string{"all"})
-				return map[string]any{"events": applyExpanded(expanded, true)}, nil
-			case "default":
-				return map[string]any{"events": traceevent.ResetToDefaultEvents()}, nil
-			case "off", "none":
-				traceevent.DisableAllEvents()
-				return map[string]any{"events": []string{}}, nil
-			case "enable":
-				if len(normalized) == 1 {
-					return nil, fmt.Errorf("trace-events enable requires at least one selector")
-				}
-				expanded, unknown := traceevent.ExpandEventSelectors(normalized[1:])
-				if len(unknown) > 0 {
-					return nil, fmt.Errorf("unknown trace events/selectors: %s", strings.Join(unknown, ", "))
-				}
-				return map[string]any{"events": applyExpanded(expanded, false)}, nil
-			case "disable":
-				if len(normalized) == 1 {
-					return nil, fmt.Errorf("trace-events disable requires at least one selector")
-				}
-				expanded, unknown := traceevent.ExpandEventSelectors(normalized[1:])
-				if len(unknown) > 0 {
-					return nil, fmt.Errorf("unknown trace events/selectors: %s", strings.Join(unknown, ", "))
-				}
-				for _, eventName := range expanded {
-					traceevent.DisableEvent(eventName)
-				}
-				return map[string]any{"events": traceevent.GetEnabledEvents()}, nil
-			default:
-				expanded, unknown := traceevent.ExpandEventSelectors(normalized)
-				if len(unknown) > 0 {
-					return nil, fmt.Errorf("unknown trace events/selectors: %s", strings.Join(unknown, ", "))
-				}
-				return map[string]any{"events": applyExpanded(expanded, true)}, nil
-			}
-
-		default:
-			return nil, fmt.Errorf("unknown setting: %s (use /set help for available settings)", key)
-		}
-	})
-
-	server.RegisterHiddenSlash("get_workflow_status", "Get workflow task status (internal)", func(args string) (any, error) {
-		slog.Info("Received get_workflow_status")
-		status, err := getWorkflowStatus(ws.GetCWD())
-		if err != nil {
-			return nil, err
-		}
-		if status == nil {
-			return nil, nil
-		}
-		return status, nil
-	})
-
-	// cycle_thinking_level removed: use /set thinking-level instead
-
-	server.RegisterHiddenSlash("get_last_assistant_text", "Get the last assistant text response (internal)", func(args string) (any, error) {
-		slog.Info("Received get_last_assistant_text")
-		messages := ag.GetMessages()
-		for i := len(messages) - 1; i >= 0; i-- {
-			if messages[i].Role == "assistant" {
-				return map[string]any{"text": messages[i].ExtractText()}, nil
-			}
-		}
-		return "", nil
-	})
-
-	server.RegisterHiddenSlash("get_fork_messages", "Get messages for a fork point (internal)", func(args string) (any, error) {
-		slog.Info("Received get_fork_messages")
-		forkMessages := sess.GetUserMessagesForForking()
-		result := make([]rpc.ForkMessage, 0, len(forkMessages))
-		for _, msg := range forkMessages {
-			result = append(result, rpc.ForkMessage{
-				EntryID: msg.EntryID,
-				Text:    msg.Text,
-			})
-		}
-		return map[string]any{"messages": result}, nil
-	})
-
-	// get_tree kept for win client compatibility (tree display still works in win)
-	server.RegisterHiddenSlash("get_tree", "Get the conversation tree structure (internal)", func(args string) (any, error) {
-		slog.Info("Received get_tree")
-		entries := sess.GetEntries()
-		tree := buildTreeEntries(entries, sess.GetLeafID())
-		return map[string]any{"entries": tree}, nil
-	})
-
-	server.RegisterSlash("rewind", "Resume generation on a specific branch", func(args string) (any, error) {
-		var jsonData struct {
-			EntryID string `json:"entryId"`
-		}
-		entryID := strings.TrimSpace(args)
-		if parseJSONArgs(args, &jsonData) && jsonData.EntryID != "" {
-			entryID = jsonData.EntryID
-		}
-		slog.Info("Received rewind", "entryId", entryID)
-		stateMu.Lock()
-		streaming := isStreaming
-		stateMu.Unlock()
-		if streaming {
-			return nil, fmt.Errorf("agent is busy")
-		}
-
-		if entryID == "" {
-			return nil, fmt.Errorf("entryId is required")
-		}
-
-		if entryID == "root" {
-			sess.ResetLeaf()
-		} else {
-			if err := sess.Branch(entryID); err != nil {
-				return nil, err
-			}
-		}
-
-		setAgentContext(createBaseContext())
-
-		// Restore llm context from the latest compaction summary on this branch
-		restoreLLMContextFromCompaction(sess)
-
-		// Update checkpoint manager (session might have changed due to branch switch)
-		if err := updateCheckpointManager(); err != nil {
-			slog.Warn("Failed to update checkpoint manager for branch resume", "error", err)
-		}
-
-		if err := sessionMgr.SaveCurrent(); err != nil {
-			slog.Info("Failed to update session metadata:", "value", err)
-		}
-		return map[string]any{"switched": true}, nil
-	})
-
-	server.RegisterSlash("fork", "Fork the conversation at a specific entry point", func(args string) (any, error) {
-		var jsonData struct {
-			EntryID string `json:"entryId"`
-		}
-		entryID := strings.TrimSpace(args)
-		if parseJSONArgs(args, &jsonData) && jsonData.EntryID != "" {
-			entryID = jsonData.EntryID
-		}
-		slog.Info("Received fork: entryId=", "value", entryID)
-		entry, ok := sess.GetEntry(entryID)
-		if !ok || entry.Type != session.EntryTypeMessage || entry.Message == nil || entry.Message.Role != "user" {
-			return nil, fmt.Errorf("invalid entryId: %s", entryID)
-		}
-
-		text := entry.Message.ExtractText()
-		name := fmt.Sprintf("fork-%s", time.Now().Format("20060102-150405"))
-		title := "Forked Session"
-		newSess, err := sessionMgr.ForkSessionFrom(sess, entry.ParentID, name, title)
-		if err != nil {
-			return nil, err
-		}
-
-		newSessionID := newSess.GetID()
-		if err := sessionMgr.SetCurrent(newSessionID); err != nil {
-			return nil, err
-		}
-		if err := sessionMgr.SaveCurrent(); err != nil {
-			slog.Info("Failed to update session metadata:", "value", err)
-		}
-
-		sess = newSess
-		sessionComp.Update(sess, compactor)
-		setAgentContext(createBaseContext())
-
-		// Update checkpoint manager for new session
-		if err := updateCheckpointManager(); err != nil {
-			slog.Warn("Failed to update checkpoint manager for fork", "error", err)
-		}
-
-		stateMu.Lock()
-		sessionID = newSessionID
-		sessionName = name
-		stateMu.Unlock()
-
-		slog.Info("Forked to new session", "name", name, "id", newSessionID)
-		return &rpc.ForkResult{Cancelled: false, Text: text}, nil
-	})
-
-	// --- Slash command aliases (user-facing short names) ---
-	// These map the short command names users type (e.g. /help, /session)
-	// to the canonical RPC handlers registered above.
-
-	// Hidden aliases: callable via RPC but not shown in /help.
-	registerHiddenAlias := func(alias, desc, canonical string) {
-		server.RegisterHiddenSlash(alias, desc, func(args string) (any, error) {
-			h, ok := server.GetSlashHandler(canonical)
-			if !ok {
-				return nil, fmt.Errorf("unknown command: /%s", canonical)
-			}
-			return h(args)
+	for _, s := range skillResult.Skills {
+		app.skillCommands = append(app.skillCommands, rpc.SlashCommand{
+			Name:        "/skill:" + s.Name,
+			Description: s.Description,
 		})
 	}
 
-	// /help — list all registered slash commands
-	server.RegisterSlash("help", "Show available slash commands", func(args string) (any, error) {
-		commands := server.ListSlashCommands()
-		return map[string]any{
-			"commands": commands,
-		}, nil
-	})
+	// --- Start event emitter ---
+	shutdownEmitter, eventEmitterDone := app.initEventEmitter()
 
-	// /skills — now the primary command (get_commands removed)
-	server.RegisterSlash("skills", "List available skills", func(args string) (any, error) {
-		slog.Info("Received skills")
-		return map[string]any{"commands": skillCommands}, nil
-	})
+	// --- Emit start event ---
+	app.emitStartEvent()
 
-	// /session — now the primary command (get_state removed)
-	// /new — now the primary command (new_session removed)
-	// /messages — now the primary command (get_messages removed)
-	// /tree — removed (broken, will be reimplemented later)
+	// --- Start debug server if enabled ---
+	app.startDebugServer()
 
-	// /thinking → shortcut for /set thinking-level
-	server.RegisterSlash("thinking", "Set thinking level (off/low/medium/high)", func(args string) (any, error) {
-		h, ok := server.GetSlashHandler("set")
-		if !ok {
-			return nil, fmt.Errorf("unknown command: set")
-		}
-		return h("thinking-level " + args)
-	})
-
-	// /trace-events → shortcut for /set trace-events
-	server.RegisterSlash("trace-events", "Configure trace events", func(args string) (any, error) {
-		h, ok := server.GetSlashHandler("set")
-		if !ok {
-			return nil, fmt.Errorf("unknown command: set")
-		}
-		return h("trace-events " + args)
-	})
-
-	// /model-select → same logic as /model (hidden, use /model instead)
-	registerHiddenAlias("model-select", "Select a model (alias for /model)", "model")
-
-	// Backward-compatible aliases for win client (internal RPC names).
-	// These old names are still sent by the win client and must be kept as aliases.
-	registerHiddenAlias("get_available_models", "List all available models (internal)", "model")
-	registerHiddenAlias("get_messages", "Get session messages (internal)", "messages")
-	registerHiddenAlias("get_state", "Get agent state (internal)", "session")
-	registerHiddenAlias("get_commands", "List commands (internal)", "skills")
-		// get_session_stats: compute real SessionStats (restored from rpc_commands.go handleGetSessionStats).
-	// Was incorrectly aliased to /session which returns SessionState, causing all-zero stats.
-	server.RegisterHiddenSlash("get_session_stats", "Get session stats (internal)", func(args string) (any, error) {
-		messages := ag.GetMessages()
-		userCount, assistantCount, toolCalls, toolResults, tokens, cost := collectSessionUsage(messages)
-
-		agentCtx := ag.GetContext()
-		tokens.SystemPromptTokens = len(systemPrompt) / 4
-		tokens.SystemToolsTokens = agentCtx.EstimateToolsTokens()
-		activeWindowTokens := agent.EstimateConversationTokens(messages)
-		tokens.ActiveWindowTokens = activeWindowTokens + tokens.SystemPromptTokens + tokens.SystemToolsTokens
-
-		var tokenRate *rpc.TokenRateStats
-		if metrics := ag.GetMetrics(); metrics != nil {
-			llmMetrics := metrics.GetLLMMetrics()
-			tokenRate = &rpc.TokenRateStats{
-				ActiveInputPerSec:   llmMetrics.ActiveInputTokensPerSec,
-				ActiveOutputPerSec:  llmMetrics.ActiveOutputTokensPerSec,
-				ActiveTotalPerSec:   llmMetrics.ActiveTotalTokensPerSec,
-				WallInputPerSec:     llmMetrics.WallInputTokensPerSec,
-				WallOutputPerSec:    llmMetrics.WallOutputTokensPerSec,
-				WallTotalPerSec:     llmMetrics.WallTotalTokensPerSec,
-				LastInputPerSec:     llmMetrics.LastInputTokensPerSec,
-				LastOutputPerSec:    llmMetrics.LastOutputTokensPerSec,
-				LastTotalPerSec:     llmMetrics.LastTotalTokensPerSec,
-				RecentWindowSeconds: llmMetrics.RecentWindowSeconds,
-				RecentInputPerSec:   llmMetrics.RecentInputTokensPerSec,
-				RecentOutputPerSec:  llmMetrics.RecentOutputTokensPerSec,
-				RecentTotalPerSec:   llmMetrics.RecentTotalTokensPerSec,
-			}
-		}
-		return &rpc.SessionStats{
-			SessionFile:       sess.GetPath(),
-			SessionID:         sessionID,
-			UserMessages:      userCount,
-			AssistantMessages: assistantCount,
-			ToolCalls:         toolCalls,
-			ToolResults:       toolResults,
-			TotalMessages:     len(messages),
-			CompactionCount:   sess.GetCompactionCount(),
-			Tokens:            tokens,
-			TokenRate:         tokenRate,
-			Cost:              cost,
-			Workspace:         ws.GetGitRoot(),
-			CurrentWorkdir:    ws.GetCWD(),
-		}, nil
-	})
-	registerHiddenAlias("new_session", "Create new session (internal)", "new")
-	registerHiddenAlias("list_sessions", "List sessions (internal)", "resume")
-	registerHiddenAlias("switch_session", "Switch session (internal)", "resume")
-
-	// For set_* commands, the win client sends args like {"enabled":true} but /set expects "subcommand value".
-	// We need custom forwarders that prepend the subcommand prefix.
-	forwardToSet := func(subcmd string) func(string) (any, error) {
-		return func(args string) (any, error) {
-			h, ok := server.GetSlashHandler("set")
-			if !ok {
-				return nil, fmt.Errorf("unknown command: set")
-			}
-			return h(subcmd + " " + args)
-		}
-	}
-	server.RegisterHiddenSlash("set_auto_compaction", "Set auto-compaction (internal)", forwardToSet("auto-compaction"))
-	server.RegisterHiddenSlash("set_thinking_level", "Set thinking level (internal)", forwardToSet("thinking-level"))
-	server.RegisterHiddenSlash("set_trace_events", "Set trace events (internal)", forwardToSet("trace-events"))
-	server.RegisterHiddenSlash("get_trace_events", "Get trace events (internal)", forwardToSet("trace-events"))
-
-	// /model — no args: list models and mark current; with args: select model by index or id
-	server.RegisterSlash("model", "List models or set the active model", func(args string) (any, error) {
-		arg := strings.TrimSpace(args)
-		if arg == "" {
-			specs, modelsPath, err := loadModelSpecs(cfg)
-			if err != nil {
-				return nil, fmt.Errorf("load models from %s: %w", modelsPath, err)
-			}
-			specs = filterModelSpecsWithKeys(specs)
-			if len(specs) == 0 {
-				authPath, _ := config.GetDefaultAuthPath()
-				return nil, fmt.Errorf("no models available (missing API keys?). Set provider keys or update %s", authPath)
-			}
-
-			models := make([]rpc.ModelInfo, 0, len(specs))
-			currentIndex := -1
-			for i, spec := range specs {
-				models = append(models, modelInfoFromSpec(spec))
-				if spec.Provider == model.Provider && spec.ID == model.ID {
-					currentIndex = i
-				}
-			}
-
-			return map[string]any{
-				"models":       models,
-				"currentIndex": currentIndex,
-				"current": map[string]any{
-					"provider": model.Provider,
-					"id":       model.ID,
-				},
-			}, nil
-		}
-
-		if idx, err := strconv.Atoi(arg); err == nil {
-			specs, modelsPath, err := loadModelSpecs(cfg)
-			if err != nil {
-				return nil, fmt.Errorf("load models from %s: %w", modelsPath, err)
-			}
-			specs = filterModelSpecsWithKeys(specs)
-			if len(specs) == 0 {
-				authPath, _ := config.GetDefaultAuthPath()
-				return nil, fmt.Errorf("no models available (missing API keys?). Set provider keys or update %s", authPath)
-			}
-			if idx < 0 || idx >= len(specs) {
-				return nil, fmt.Errorf("model index out of range")
-			}
-			arg = fmt.Sprintf("%s %s", specs[idx].Provider, specs[idx].ID)
-		} else if strings.Contains(arg, "/") {
-			parts := strings.SplitN(arg, "/", 2)
-			provider := strings.TrimSpace(parts[0])
-			modelID := strings.TrimSpace(parts[1])
-			if provider == "" || modelID == "" {
-				return nil, fmt.Errorf("invalid model selector: %s", arg)
-			}
-			arg = fmt.Sprintf("%s %s", provider, modelID)
-		}
-
-		h, ok := server.GetSlashHandler("set_model")
-		if !ok {
-			return nil, fmt.Errorf("unknown command: set_model")
-		}
-		return h(arg)
-	})
-
-	// /resume — no args: list sessions; with args: switch to session
-	server.RegisterSlash("resume", "List sessions or resume a session by ID/name", func(args string) (any, error) {
-		arg := strings.TrimSpace(args)
-		if arg == "" {
-			// List all sessions directly (was previously delegating to list_sessions alias → infinite recursion)
-			sessions, err := sessionMgr.ListSessions()
-			if err != nil {
-				return nil, fmt.Errorf("failed to list sessions: %w", err)
-			}
-			return map[string]any{"sessions": sessions}, nil
-		}
-
-		// Resolve session by index, ID, or name
-		var targetID string
-		sessions, err := sessionMgr.ListSessions()
-		if err != nil {
-			return nil, fmt.Errorf("failed to list sessions: %w", err)
-		}
-
-		// Try numeric index first
-		if idx, err := strconv.Atoi(arg); err == nil {
-			if idx < 0 || idx >= len(sessions) {
-				return nil, fmt.Errorf("session index %d out of range (0-%d)", idx, len(sessions)-1)
-			}
-			targetID = sessions[idx].ID
-		} else {
-			// Try exact ID match, then name match
-			for _, s := range sessions {
-				if s.ID == arg {
-					targetID = s.ID
-					break
-				}
-			}
-			if targetID == "" {
-				for _, s := range sessions {
-					if s.Name == arg {
-						targetID = s.ID
-						break
-					}
-				}
-			}
-			if targetID == "" {
-				return nil, fmt.Errorf("session not found: %s", arg)
-			}
-		}
-
-		// Load the target session
-		newSess, err := sessionMgr.GetSession(targetID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load session %s: %w", targetID, err)
-		}
-
-		// Update session manager's current ID
-		if err := sessionMgr.SetCurrent(targetID); err != nil {
-			return nil, err
-		}
-		if err := sessionMgr.SaveCurrent(); err != nil {
-			slog.Info("Failed to update session metadata:", "value", err)
-		}
-
-		// Switch to the new session
-		newSessionName := resolveSessionName(sessionMgr, targetID)
-		sess = newSess
-		sessionComp.Update(sess, compactor)
-		setAgentContext(createBaseContext())
-
-		// Update checkpoint manager for switched session
-		if err := updateCheckpointManager(); err != nil {
-			slog.Warn("Failed to update checkpoint manager for session switch", "error", err)
-		}
-
-		stateMu.Lock()
-		sessionID = targetID
-		sessionName = newSessionName
-		stateMu.Unlock()
-
-		slog.Info("Switched to session", "id", targetID, "name", newSessionName)
-		server.EmitEvent(map[string]any{"type": "session_switch", "session": targetID, "sessionName": newSessionName})
-		return map[string]any{"sessionId": targetID, "sessionName": newSessionName}, nil
-	})
-
-	// /context — composite: returns state + session stats + available models
-	server.RegisterSlash("context", "Show current state, session stats, and available models", func(args string) (any, error) {
-		stateH, _ := server.GetSlashHandler("get_state")
-		statsH, _ := server.GetSlashHandler("get_session_stats")
-		modelsH, _ := server.GetSlashHandler("get_available_models")
-
-		stateResult, err := stateH("")
-		if err != nil {
-			return nil, err
-		}
-		statsResult, _ := statsH("")
-		modelsResult, _ := modelsH("")
-
-		return map[string]any{
-			"state":  stateResult,
-			"stats":  statsResult,
-			"models": modelsResult,
-		}, nil
-	})
-
-	// Display settings: /toggle thinking|prefix, /set tools|busy-mode
-	server.RegisterSlash("toggle", "Toggle display settings (thinking, prefix, tools)", func(args string) (any, error) {
-		kind := strings.TrimSpace(args)
-		switch kind {
-		case "thinking":
-			showThinking = !showThinking
-			return map[string]any{"setting": "thinking", "value": showThinking}, nil
-		case "prefix":
-			showPrefix = !showPrefix
-			return map[string]any{"setting": "prefix", "value": showPrefix}, nil
-		case "tools":
-			showTools = !showTools
-			return map[string]any{"setting": "tools", "value": showTools}, nil
-		default:
-			return nil, fmt.Errorf("usage: /toggle <thinking|prefix|tools>")
-		}
-	})
-
-	server.RegisterSlash("show", "Show agent settings or pipeline info", func(args string) (any, error) {
-		subCmd := strings.TrimSpace(args)
-		switch subCmd {
-		case "settings", "":
-			compaction := buildCompactionState(compactorConfig, compactor)
-
-			model := currentModelInfo.ID
-			if currentModelInfo.Provider != "" {
-				model = currentModelInfo.Provider + "/" + currentModelInfo.ID
-			}
-
-			compactionContext := "unknown"
-			compactionReserve := "unknown"
-			compactionLimit := "unknown"
-			compactionMaxMessages := "disabled"
-			compactionMaxTokens := "disabled"
-			compactionKeepRecent := "unknown"
-			compactionKeepRecentTokens := "unknown"
-			if compaction != nil {
-				compactionContext = formatIntOrUnknown(compaction.ContextWindow)
-				compactionReserve = formatIntOrUnknown(compaction.ReserveTokens)
-				compactionLimit = formatTokenLimit(compaction)
-				compactionMaxMessages = formatLimit(compaction.MaxMessages)
-				compactionMaxTokens = formatLimit(compaction.MaxTokens)
-				compactionKeepRecent = formatIntOrUnknown(compaction.KeepRecent)
-				compactionKeepRecentTokens = formatIntOrUnknown(compaction.KeepRecentTokens)
-			}
-
-			autoCompStr := "off"
-			if autoCompactionEnabled {
-				autoCompStr = "on"
-			}
-
-			showThinkingStr := "off"
-			if showThinking {
-				showThinkingStr = "on"
-			}
-			showToolsStr := "off"
-			if showTools {
-				showToolsStr = "on"
-			}
-			showPrefixStr := "off"
-			if showPrefix {
-				showPrefixStr = "on"
-			}
-
-			return map[string]any{
-				"type": "settings",
-				"data": map[string]any{
-					"model":                         model,
-					"show-thinking":                 showThinkingStr,
-					"tools":                         showToolsStr,
-					"prefix":                        showPrefixStr,
-					"thinking-level":                currentThinkingLevel,
-					"busy-mode":                     busyMode,
-					"auto-compaction":               autoCompStr,
-					"compaction-context-window":     compactionContext,
-					"compaction-reserve-tokens":     compactionReserve,
-					"compaction-token-limit":        compactionLimit,
-					"compaction-max-messages":       compactionMaxMessages,
-					"compaction-max-tokens":         compactionMaxTokens,
-					"compaction-keep-recent":        compactionKeepRecent,
-					"compaction-keep-recent-tokens": compactionKeepRecentTokens,
-				},
-			}, nil
-		case "pipeline":
-			return map[string]any{"message": "pipeline info not yet available"}, nil
-		default:
-			return nil, fmt.Errorf("usage: /show settings|pipeline")
-		}
-	})
-
-	// /quit — quit the application
-	server.RegisterSlash("quit", "Exit the application", func(args string) (any, error) {
-		slog.Info("Received quit command, exiting application")
-		os.Exit(0)
-		return nil, nil // unreachable, but Go requires return
-	})
-
-	// /abort — abort the current agent execution
-	server.RegisterSlash("abort", "Abort the current agent execution", func(args string) (any, error) {
-		slog.Info("Received abort command")
-		stateMu.Lock()
-		streaming := isStreaming
-		stateMu.Unlock()
-
-		if !streaming {
-			return nil, fmt.Errorf("agent is not streaming")
-		}
-
-		ag.Abort()
-		return map[string]any{"status": "aborting"}, nil
-	})
-
-	// /follow-up — add a follow-up message (same as /follow_up)
-	server.RegisterSlash("follow-up", "Add a follow-up message when agent is busy", func(args string) (any, error) {
-		message := strings.TrimSpace(args)
-		if message == "" {
-			return nil, fmt.Errorf("usage: /follow-up <message>")
-		}
-
-		slog.Info("Received follow-up command")
-		stateMu.Lock()
-		streaming := isStreaming
-		stateMu.Unlock()
-
-		if !streaming {
-			return nil, fmt.Errorf("agent is not busy")
-		}
-
-		// Check if follow-up mode allows this
-		if followUpMode != "one-at-a-time" && followUpMode != "queue" {
-			return nil, fmt.Errorf("follow-up mode is '%s', not enabled", followUpMode)
-		}
-
-		// Check if there's already a pending follow-up
-		if len(followUpQueue) > 0 && followUpMode == "one-at-a-time" {
-			return nil, fmt.Errorf("follow-up queue already has a pending message")
-		}
-
-		expandedMessage := expandSkillCommands(message)
-		followUpQueue = append(followUpQueue, expandedMessage)
-		return map[string]any{"status": "queued", "message": expandedMessage}, nil
-	})
-
-	// /steer — alias for steering when agent is busy
-	// steer removed: now default behavior when agent is streaming (busy-mode default is steer)
-
-	// Start event emitter
-	eventEmitterDone := make(chan struct{})
-	shutdownEmitter := make(chan struct{})
-	go func() {
-		defer close(eventEmitterDone)
-		for {
-			select {
-			case event := <-ag.Events():
-				if event.Type == "agent_start" {
-					stateMu.Lock()
-					isStreaming = true
-					stateMu.Unlock()
-				}
-				if event.Type == "agent_end" {
-					stateMu.Lock()
-					isStreaming = false
-					isCompacting = false
-					pendingSteer = false
-					stateMu.Unlock()
-				}
-				if event.Type == "compaction_start" {
-					stateMu.Lock()
-					isCompacting = true
-					stateMu.Unlock()
-				}
-				if event.Type == "compaction_end" {
-					stateMu.Lock()
-					isCompacting = false
-					stateMu.Unlock()
-				}
-
-				if event.Type == "message_end" && event.Message != nil {
-					if sessionWriter != nil {
-						sessionWriter.Append(sess, *event.Message)
-					}
-				}
-				if event.Type == "tool_execution_end" && event.Result != nil {
-					if sessionWriter != nil {
-						sessionWriter.Append(sess, *event.Result)
-					}
-				}
-				if event.Type == "agent_end" && sessionWriter != nil {
-					// No longer Replace (full rewrite of messages.jsonl).
-					// Messages are appended one-by-one during the turn via sessionWriter.Append.
-					// Compact events are appended via AppendCompactEvent.
-				}
-
-				emitAt := time.Now()
-				if event.EventAt == 0 {
-					event.EventAt = emitAt.UnixNano()
-				}
-				server.EmitEvent(event)
-
-				if event.Type == "agent_end" {
-					go func() {
-						if err := sessionMgr.SaveCurrent(); err != nil {
-							slog.Info("Failed to update session metadata:", "value", err)
-						}
-					}()
-				}
-
-			case <-shutdownEmitter:
-				for {
-					select {
-					case event := <-ag.Events():
-						if event.Type == "agent_start" {
-							stateMu.Lock()
-							isStreaming = true
-							stateMu.Unlock()
-						}
-						if event.Type == "agent_end" {
-							stateMu.Lock()
-							isStreaming = false
-							isCompacting = false
-							pendingSteer = false
-							stateMu.Unlock()
-						}
-						if event.Type == "compaction_start" {
-							stateMu.Lock()
-							isCompacting = true
-							stateMu.Unlock()
-						}
-						if event.Type == "compaction_end" {
-							stateMu.Lock()
-							isCompacting = false
-							stateMu.Unlock()
-						}
-
-						if event.Type == "message_end" && event.Message != nil {
-							if sessionWriter != nil {
-								sessionWriter.Append(sess, *event.Message)
-							}
-						}
-						if event.Type == "tool_execution_end" && event.Result != nil {
-							if sessionWriter != nil {
-								sessionWriter.Append(sess, *event.Result)
-							}
-						}
-						if event.Type == "agent_end" && sessionWriter != nil {
-							// No longer Replace (full rewrite of messages.jsonl).
-							// Messages are appended one-by-one during the turn via sessionWriter.Append.
-							// Compact events are appended via AppendCompactEvent.
-						}
-
-						emitAt := time.Now()
-						if event.EventAt == 0 {
-							event.EventAt = emitAt.UnixNano()
-						}
-						server.EmitEvent(event)
-						if event.Type == "agent_end" {
-							go func() {
-								if err := sessionMgr.SaveCurrent(); err != nil {
-									slog.Info("Failed to update session metadata:", "value", err)
-								}
-							}()
-						}
-					default:
-						return
-					}
-				}
-			}
-		}
-	}()
-
-	// Emit start event
-	allTools := registry.All()
-	toolNames := make([]string, len(allTools))
-	for i, t := range allTools {
-		toolNames[i] = t.Name()
-	}
-	server.EmitEvent(map[string]any{
-		"type":  "server_start",
-		"model": model.ID,
-		"tools": toolNames,
-	})
-
-	// Start debug server if enabled
-	if debugAddr != "" {
-		go func() {
-			// Register metrics endpoint on DefaultServeMux
-			http.HandleFunc("/debug/metrics", func(w http.ResponseWriter, r *http.Request) {
-				metrics := ag.GetMetrics()
-				if metrics == nil {
-					http.Error(w, "Metrics not available", http.StatusServiceUnavailable)
-					return
-				}
-
-				fullMetrics := metrics.GetFullMetrics()
-				w.Header().Set("Content-Type", "application/json")
-				if err := json.NewEncoder(w).Encode(fullMetrics); err != nil {
-					slog.Error("Failed to encode metrics:", "value", err)
-					http.Error(w, "Failed to encode metrics", http.StatusInternalServerError)
-				}
-			})
-
-			slog.Info("Debug server listening on", "value", debugAddr)
-			slog.Info("Debug endpoints available at:")
-			slog.Info("- http:///debug/pprof/          (profiling index)", "value", debugAddr)
-			slog.Info("- http:///debug/pprof/profile   (CPU profile)", "value", debugAddr)
-			slog.Info("- http:///debug/pprof/heap       (memory profile)", "value", debugAddr)
-			slog.Info("- http:///debug/pprof/goroutine  (goroutine dump)", "value", debugAddr)
-			slog.Info("- http:///debug/pprof/trace      (execution trace)", "value", debugAddr)
-			slog.Info("- http:///debug/metrics         (agent metrics)", "value", debugAddr)
-
-			if err := http.ListenAndServe(debugAddr, nil); err != nil {
-				slog.Error("Debug server error:", "error", err)
-			}
-		}()
-	}
-
-	// Run RPC server
+	// --- Run RPC server ---
 	slog.Info("RPC server started", "model", model.ID, "cwd", cwd)
 	slog.Info("Waiting for commands...")
 	runErr := server.RunWithIO(input, output)
